requirements.txt
Django==4.2.10
whitenoise==6.5.0
python-dotenv 